# docs/openapi/OpenAPI Reference Implementation.json
{
  "swagger": "2.0",
  "info": {
    "title": "Identfy API",
    "description": "This api is Identfy",
    "license": {
      "name": "MIT License"
    },
    "version": "v1"
  },
  "schemes": [
    "https"
  ],
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "basic": {
      "type": "basic"
    }
  },
  "security": [
    {
      "Bearer": []
    },
    {
      "basic": []
    }
  ],
  "paths": {
...
    "/credential-offer/": {
      "post": {
        "operationId": "resolve_credential_offer",
        "description": "POST CredentialOffer",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ResolveCredentialOffer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ResolveCredentialOfferResponse"
            }
          }
        },
        "tags": [
          "resolve_credential_offer"
        ]
      },
      "parameters": []
    },
...
    "/credential_deferred/": {
      "post": {
        "operationId": "deferred_credentials",
        "description": "POST Deferred Credentials",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Auth Token",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/CredentialResponse"
            }
          }
        },
        "tags": [
          "deferred_credentials"
        ]
      },
      "parameters": []
    },
    "/credentials/": {
      "post": {
        "operationId": "credentials",
        "description": "POST Credentials",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EbsiCredentialRequest"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Auth Token",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/CredentialResponse"
            }
          }
        },
        "tags": [
          "credentials"
        ]
      },
      "parameters": []
    },
    "/credentials/ebsi/accreditation/": {
      "get": {
        "operationId": "credentials_ebsi_ebsi_accreditation_direct_issuance",
        "description": "GET VC Status",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "The accreditation type to issue",
            "required": true,
            "type": "string"
          },
          {
            "name": "holder",
            "in": "query",
            "description": "The DID of the holder",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "credentials"
        ]
      },
      "parameters": []
    },
    "/credentials/external-data/": {
      "get": {
        "operationId": "credentials_external_data_credentials",
        "description": "GET Credential External Data",
        "parameters": [
          {
            "name": "vc_type",
            "in": "query",
            "description": "Credential Schema Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "User identifier",
            "required": false,
            "type": "string"
          },
          {
            "name": "pin",
            "in": "query",
            "description": "User identifier",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ExternalDataResponse"
            }
          }
        },
        "tags": [
          "credentials"
        ]
      },
      "parameters": []
    },
    "/credentials/request-deferred/": {
      "post": {
        "operationId": "credentials_exchange_deferred_vc",
        "description": "POST Request Deferred",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestDeferredVc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/RequestVcResponse"
            }
          }
        },
        "tags": [
          "credentials"
        ]
      },
      "parameters": []
    },
    "/credentials/request/": {
      "post": {
        "operationId": "credentials_request_vc",
        "description": "POST Request VC",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestVc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/RequestVcResponse"
            }
          }
        },
        "tags": [
          "credentials"
        ]
      },
      "parameters": []
    },
...
    "/deferred/register/": {
      "post": {
        "operationId": "deferred_register_deferred",
        "description": "POST Registry Deferred token",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeferredRegistry"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "deferred"
        ]
      },
      "parameters": []
    },
...
  },
  "definitions": {
...
    "ResolveCredentialOffer": {
      "required": [
        "credential_offer"
      ],
      "type": "object",
      "properties": {
        "credential_offer": {
          "title": "Credential offer",
          "description": "Credential Offer from a VC Issuer",
          "type": "string",
          "minLength": 1
        }
      }
    },
    "ResolveCredentialOfferResponse": {
      "type": "object",
      "properties": {}
    },
...
    "CredentialResponse": {
      "type": "object",
      "properties": {
        "format": {
          "title": "Format",
          "type": "string",
          "default": "jwt_vc",
          "maxLength": 255,
          "minLength": 1
        },
        "credential": {
          "title": "Credential",
          "type": "string",
          "maxLength": 1255,
          "minLength": 1
        },
        "acceptance_token": {
          "title": "Acceptance token",
          "type": "string",
          "maxLength": 1255,
          "minLength": 1
        },
        "c_nonce": {
          "title": "C nonce",
          "type": "string",
          "maxLength": 255,
          "minLength": 1
        },
        "c_nonce_expires_in": {
          "title": "C nonce expires in",
          "type": "integer"
        }
      }
    },
    "EbsiCredentialRequest": {
      "required": [
        "types",
        "proof"
      ],
      "type": "object",
      "properties": {
        "types": {
          "title": "Types",
          "description": "Credential types requested",
          "type": "object"
        },
        "format": {
          "title": "Format",
          "description": "Format of the returned credentials",
          "type": "string",
          "default": "jwt_vc",
          "maxLength": 255,
          "minLength": 1
        },
        "proof": {
          "title": "Proof",
          "description": "Proof object containing proof_type and jwt",
          "type": "object"
        }
      }
    },
    "ExternalDataResponse": {
      "type": "object",
      "properties": {
        "credential_data": {
          "title": "Credential data",
          "description": "Credential data",
          "type": "object"
        },
        "schema_addr": {
          "title": "Schema addr",
          "type": "string",
          "maxLength": 1255,
          "minLength": 1
        }
      }
    },
    "RequestDeferredVc": {
      "required": [
        "issuer",
        "acceptanceToken"
      ],
      "type": "object",
      "properties": {
        "issuer": {
          "title": "Issuer",
          "description": "VC Issuer URI",
          "type": "string",
          "minLength": 1
        },
        "acceptanceToken": {
          "title": "Acceptancetoken",
          "description": "Token to exchange for the VC",
          "type": "string",
          "minLength": 1
        }
      }
    },
    "RequestVcResponse": {
      "type": "object",
      "properties": {
        "credential": {
          "title": "Credential",
          "description": "Credential Offer from a VC Issuer",
          "type": "string",
          "minLength": 1
        },
        "acceptance_token": {
          "title": "Acceptance token",
          "description": "Exchange Code for a Deferred VC",
          "type": "string",
          "minLength": 1
        }
      }
    },
    "RequestVc": {
      "required": [
        "credential_offer",
        "vc_type"
      ],
      "type": "object",
      "properties": {
        "credential_offer": {
          "title": "Credential offer",
          "description": "Credential Offer from a VC Issuer",
          "type": "string",
          "minLength": 1
        },
        "vc_type": {
          "description": "Types of the VC to request",
          "type": "array",
          "items": {
            "type": "string",
            "minLength": 1
          }
        },
        "pin_code": {
          "title": "Pin code",
          "type": "integer"
        }
      }
    },
...
    "DeferredRegistry": {
      "required": [
        "client_id",
        "vc_type"
      ],
      "type": "object",
      "properties": {
        "client_id": {
          "title": "Client id",
          "type": "string",
          "maxLength": 255,
          "minLength": 1
        },
        "vc_type": {
          "title": "Vc type",
          "type": "string",
          "maxLength": 255,
          "minLength": 1
        },
        "pin": {
          "title": "Pin",
          "type": "string",
          "maxLength": 255,
          "minLength": 1
        }
      }
    },
...
  }
}